backend/requirements.txt
Flask==2.3.3
flask-cors==3.0.10
gunicorn==21.2.0
PyJWT>=2.7.0
